# rpi5_snapshot/openapi.yaml
openapi: 3.0.3
info:
  title: Game Stats API
  version: 1.0.0
  description: API kontrakty dla modułów 1v1 oraz multiplayer.
servers:
  - url: /api/v1
components:
  schemas:
    Uuid:
      type: string
      format: uuid
      example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
    DateTime:
      type: string
      format: date-time
      example: "2026-01-30T12:00:00.000Z"
    ErrorDetailsItem:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    ErrorPayload:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR]
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetailsItem'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
...
    Game:
      type: object
      required: [id, code, name, isActive]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        name:
          type: string
        isActive:
          type: boolean
    GameDeleteResponse:
      type: object
      required: [id, code, name, deletedMatches]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        code:
          type: string
        name:
          type: string
        deletedMatches:
          type: integer
    Player:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
        isActive:
          type: boolean
        createdAt:
          $ref: '#/components/schemas/DateTime'
    PlayerDeleteResponse:
      type: object
      required:
        [
          player,
          deletedOneVsOneMatches,
          deletedMultiplayerParticipations,
          deletedOrphanMultiplayerMatches,
        ]
      properties:
        player:
          $ref: '#/components/schemas/Player'
        deletedOneVsOneMatches:
          type: integer
        deletedMultiplayerParticipations:
          type: integer
        deletedOrphanMultiplayerMatches:
          type: integer
...
    PlayerCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
    PlayerPatch:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
        isActive:
          type: boolean
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /games:
    get:
      summary: List 1v1 games
      parameters:
        - in: query
          name: includeInactive
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /games/{id}:
    patch:
      summary: Update 1v1 game active flag
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive:
                  type: boolean
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete 1v1 game with related matches
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDeleteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /players:
    get:
      summary: List players
      parameters:
        - in: query
          name: active
          schema:
            type: boolean
            default: true
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /players/{id}:
    patch:
      summary: Update player
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerPatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      summary: Delete player with related matches and stats
      parameters:
        - in: path
          name: id
          required: true
          schema:
            $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDeleteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
...